+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.1] - 2022-07-11
+### Changed
+
+- Move & transform endpoint logo to generic logo in the payment methods tab
+- Add a refresh link when iframe payment is rejected
+
 ## [1.2.0] - 2022-04-12
 ### Added